Adding exception handling for invalid ID into TeamServiceImpl class

+++ src/main/java/com/felipe/taskmanagementeapi/services/impl/TeamServiceImpl.java
@@ -3,6 +3,7 @@ package com.felipe.taskmanagementeapi.services.impl;
 import com.felipe.taskmanagementeapi.Repositories.TeamRepository;
 import com.felipe.taskmanagementeapi.dtos.TeamDto;
 import com.felipe.taskmanagementeapi.entities.TeamEntity;
+import com.felipe.taskmanagementeapi.exceptions.ResourceNotFoundException;
 import com.felipe.taskmanagementeapi.services.TeamService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,6 +30,10 @@ public class TeamServiceImpl implements TeamService {
     @Override
     public TeamDto findTeamById(Integer id) {
         Optional<TeamEntity> savedTeam = teamRepository.findById(id);
+        if(savedTeam.isEmpty()) {
+            throw new ResourceNotFoundException("Team", "id", id);
+        }
+
         TeamDto teamDto = new TeamDto();
         BeanUtils.copyProperties(savedTeam.get(), teamDto);
         return teamDto;
@@ -44,6 +49,10 @@ public class TeamServiceImpl implements TeamService {
     @Override
     public TeamDto updateTeamName(TeamDto teamDto, Integer id) {
         Optional<TeamEntity> savedTeam = teamRepository.findById(id);
+        if(savedTeam.isEmpty()) {
+            throw new ResourceNotFoundException("Team", "id", id);
+        }
+
         savedTeam.get().setName(teamDto.getName());
         teamRepository.save(savedTeam.get());
         BeanUtils.copyProperties(savedTeam.get(), teamDto);
@@ -53,6 +62,9 @@ public class TeamServiceImpl implements TeamService {
     @Override
     public String deleteTeamById(Integer id) {
         Optional<TeamEntity> savedTeam = teamRepository.findById(id);
+        if(savedTeam.isEmpty()) {
+            throw new ResourceNotFoundException("Team", "id", id);
+        }
         teamRepository.deleteById(id);
         return "Team successfully deleted!";
     }
